Implement versioned research execution and strategy composition

+++ src/scheme/models.py
@@ -4,4 +4,4 @@ __all__ = ["Asset", "Order", "Signal", "Target"]
 
 type Asset = str
 type Signal[T] = dict[Asset, T]
-type Target = dict[Asset, int]
+type Target = dict[Asset, float]
